Use compenstated sum in order parameter statistics
