--- src/field_stats_q.h
@@ -1,23 +1,0 @@
-/*****************************************************************************
- *
- *  field_stats_q.h
- *
- *
- *  Edinburgh Soft Matter and Statistical Physics Group and
- *  Edinburgh Parallel Computing Centre
- *
- *  (c) 2022 The University of Edinburgh
- *
- *  Kevin Stratford (kevin@epcc.ed.ac.uk)
- *
- *****************************************************************************/
-
-#ifndef LUDWIG_FIELD_STATS_Q_H
-#define LUDWIG_FIELD_STATS_Q_H
-
-#include "field.h"
-#include "map.h"
-
-int field_stats_q(field_t * q, map_t * map);
-
-#endif
