+++ tests/regression/d3q19-short/serial-chol-n02.log
@@ -166,7 +166,7 @@ Scalars - total mean variance min max
 [phi] -2.3050805e-15 -8.9001309e-21  2.6690464e-08 -7.0823033e-03  7.0823033e-03
 [phi]  4.8169716e-15  1.8598777e-20  3.3769091e-08 -7.5924269e-03  7.5924269e-03
 [phi]  2.1265219e+04  8.2106995e-02  3.1278470e-02 -1.6690066e-01  3.3397530e-01
-[phi]  1.8653835e-02  7.2024197e-08  3.1212676e-02 -2.5040482e-01  2.5040481e-01
+[phi]  1.8653836e-02  7.2024203e-08  3.1212676e-02 -2.5040482e-01  2.5040481e-01
 
 Free energies - timestep f v f/v f_s a f_s/a
 [fe]              5 -1.6870562224e+01  2.5899400000e+05 -6.5138814893e-05  8.4425272198e-01  2.0280000000e+03  4.1629818638e-04
