+++ tests/regression/d3q19-short/serial-drop-lc3.log
@@ -121,11 +121,11 @@ Initial conditions.
 Scalars - total mean variance min max
 [rho]      262144.00  1.00000000000  2.2204460e-16  1.00000000000  1.00000000000
 [phi]  2.1068800e+05  8.0371094e-01 3.5404873e-01 -1.0000000e+00 1.0000000e+00
-[phi]  4.1760137e-12  1.5930228e-17 6.0000000e-02 -4.4966062e-01 4.4966064e-01
-[phi] -3.1326420e-08 -1.1950081e-13 5.0000000e-02 -6.0000000e-01 6.0000000e-01
-[phi] -3.1329084e-08 -1.1951097e-13 5.0000000e-02 -6.0000000e-01 6.0000000e-01
-[phi]  1.4539411e-13  5.5463453e-19 6.0000000e-02 -4.4966062e-01 4.4966064e-01
-[phi] -3.1333445e-08 -1.1952761e-13 5.0000000e-02 -6.0000000e-01 6.0000000e-01
+[phi] -4.7213797e-14 -1.8010634e-19 6.0000000e-02 -4.4966062e-01 4.4966064e-01
+[phi] -3.1328486e-08 -1.1950869e-13 5.0000000e-02 -6.0000000e-01 6.0000000e-01
+[phi] -3.1328486e-08 -1.1950869e-13 5.0000000e-02 -6.0000000e-01 6.0000000e-01
+[phi] -4.7213797e-14 -1.8010634e-19 6.0000000e-02 -4.4966062e-01 4.4966064e-01
+[phi] -3.1328486e-08 -1.1950869e-13 5.0000000e-02 -6.0000000e-01 6.0000000e-01
 
 Momentum - x y z
 [total   ]  0.0000000e+00  0.0000000e+00  0.0000000e+00
