+++ tests/regression/d3q19-short/serial-drop-lc5.log
@@ -115,10 +115,10 @@ Initial conditions.
 Scalars - total mean variance min max
 [rho]       32768.00  1.00000000000  2.2204460e-16  1.00000000000  1.00000000000
 [phi]  1.8176000e+04  5.5468750e-01 6.9232178e-01 -1.0000000e+00 1.0000000e+00
-[phi]  4.0040471e-12  1.2219382e-16 6.0000000e-02 -4.4966055e-01 4.4966059e-01
+[phi] -8.6591172e-15 -2.6425529e-19 6.0000000e-02 -4.4966055e-01 4.4966059e-01
 [phi] -3.8895517e+03 -1.1869970e-01 3.8159319e-02 -6.0000000e-01 4.8051711e-01
 [phi] -3.8895517e+03 -1.1869970e-01 3.8159319e-02 -6.0000000e-01 4.8051711e-01
-[phi]  5.6144776e-12  1.7134026e-16 6.0000000e-02 -4.4966055e-01 4.4966059e-01
+[phi] -8.6591172e-15 -2.6425529e-19 6.0000000e-02 -4.4966055e-01 4.4966059e-01
 [phi] -3.8895517e+03 -1.1869970e-01 3.8159319e-02 -6.0000000e-01 4.8051711e-01
 
 Momentum - x y z
