+++ tests/regression/d3q19-short/serial-init-bp2.log
@@ -89,11 +89,11 @@ Initial conditions.
 
 Scalars - total mean variance min max
 [rho]       32768.00  1.00000000000  2.2204460e-16  1.00000000000  1.00000000000
-[phi]  2.6610658e-15  8.1209284e-20 8.9999982e-02 -6.0000000e-01 6.0000000e-01
+[phi]  0.0000000e+00  0.0000000e+00 8.9999982e-02 -6.0000000e-01 6.0000000e-01
 [phi] -4.0156293e-04 -1.2254728e-08 4.5000009e-02 -3.0000000e-01 3.0000000e-01
-[phi] -4.0156294e-04 -1.2254728e-08 4.5000009e-02 -3.0000000e-01 3.0000000e-01
-[phi] -8.9217488e-12 -2.7227016e-16 8.9999982e-02 -6.0000000e-01 6.0000000e-01
-[phi] -4.0156287e-04 -1.2254726e-08 4.5000009e-02 -3.0000000e-01 3.0000000e-01
+[phi] -4.0156293e-04 -1.2254728e-08 4.5000009e-02 -3.0000000e-01 3.0000000e-01
+[phi]  0.0000000e+00  0.0000000e+00 8.9999982e-02 -6.0000000e-01 6.0000000e-01
+[phi] -4.0156293e-04 -1.2254726e-08 4.5000009e-02 -3.0000000e-01 3.0000000e-01
 
 Momentum - x y z
 [total   ]  4.5474735e-13  0.0000000e+00  0.0000000e+00
@@ -103,10 +103,10 @@ Starting time step loop.
 
 Scalars - total mean variance min max
 [rho]       32768.00  1.00000000000  5.5946818e-07  0.99868639745  1.00197940975
-[phi] -2.8587896e-13 -8.7243335e-18 8.9762988e-02 -5.9884923e-01 5.9934603e-01
+[phi]  2.5125264e-15  7.6676219e-20 8.9762988e-02 -5.9884923e-01 5.9934603e-01
 [phi] -4.0099021e-04 -1.2237250e-08 4.4898274e-02 -2.9996132e-01 2.9996132e-01
 [phi] -4.0099021e-04 -1.2237250e-08 4.4898274e-02 -2.9996132e-01 2.9996132e-01
-[phi]  3.4842060e-12  1.0632953e-16 8.9762988e-02 -5.9884923e-01 5.9934603e-01
+[phi]  9.8944335e-16  3.0195415e-20 8.9762988e-02 -5.9884923e-01 5.9934603e-01
 [phi] -4.0099021e-04 -1.2237250e-08 4.4898274e-02 -2.9996132e-01 2.9996132e-01
 
 Free energies - timestep f v f/v f_bulk/v f_grad/v redshift
